+++ build-tools/rel-build.sh
@@ -4,6 +4,7 @@
 set -ex
 
 CURDIR="$(dirname $BASH_SOURCE)"
+RUN_TESTS=${RUN_TESTS:-1}
 
 . $CURDIR/_build-lib.sh
 BUILDDIR=$(get_builddir)
@@ -12,16 +13,20 @@ export BUILDDIR=$BUILDDIR
 
 go_install $(all_cmds)
 
-echo "Gathering unit test code coverage for 'release' build..."
-ginkgo_test_with_coverage
+if [ $RUN_TESTS -eq 1 ]; then
+    echo "Gathering unit test code coverage for 'release' build..."
+    ginkgo_test_with_coverage
+fi
 
 # reset GOPATH after using temp directories
 export GOPATH=/build
 
-# push coverage data to coveralls if F5 repo or if configured for fork.
-if [ "$COVERALLS_TOKEN" ]; then
-  cat $BUILDDIR/coverage/coverage.out >> $BUILDDIR/coverage.out
-  goveralls \
-    -coverprofile=$BUILDDIR/coverage.out \
-    -service=travis-ci
+if [ $RUN_TESTS -eq 1 ]; then
+    # push coverage data to coveralls if F5 repo or if configured for fork.
+    if [ "$COVERALLS_TOKEN" ]; then
+      cat $BUILDDIR/coverage/coverage.out >> $BUILDDIR/coverage.out
+      goveralls \
+        -coverprofile=$BUILDDIR/coverage.out \
+        -service=travis-ci
+    fi
 fi