+++ docs/RELEASE-NOTES.rst
@@ -4,10 +4,13 @@ Release Notes for BIG-IP Controller for Kubernetes
 next-release
 ------------
 
+v1.7.0
+------
+
 Added Functionality
 ```````````````````
 * Added `--manage-configmaps` argument to CC to prevent or allow CC to respond to ConfigMap events. Defaults to `true`.
-* Added `virtual-server.f5.com/whitelist-source-range` annotation to support CIDR whitelisting.
+* Added `virtual-server.f5.com/whitelist-source-range` Ingress/Route annotation to support IP CIDR whitelisting.
 * :issues:`699` - Ability to configure health monitor type in Ingress/Route annotation. Http is the default.
 * Changed container base image to use debian-slim.